combine two footer folders
