--- src/components/DPLFooter/index.tsx
@@ -1,21 +1,0 @@
-import React from 'react'
-import { DPLHr } from '../DPLHr'
-
-interface DPLFooterProps {
-  className?: string
-}
-
-const DPLFooter: React.FC<DPLFooterProps> = ({ className, children }) => {
-  return (
-    <footer className={`mt-lg ${className}`}>
-      <DPLHr />
-      <div className="py-md ">{children}</div>
-    </footer>
-  )
-}
-
-const DPLFooterSection: React.FC<DPLFooterProps> = ({ children }) => {
-  return <div className="grid justify-items-start gap-8">{children}</div>
-}
-
-export { DPLFooter, DPLFooterSection }
