+++ src/components/Footer/index.tsx
@@ -2,10 +2,27 @@ import React from 'react'
 import { FaDiscord, FaGithubSquare } from 'react-icons/fa'
 import { Link } from 'react-router-dom'
 import { DEPLOYMENTS } from '../../const'
-import { DPLFooter, DPLFooterSection } from '../DPLFooter'
+import { DPLHr } from '../DPLHr'
 import FooterImg from '../../img/FOOTER_IMG_Powered by Dev Protocol.svg'
 
 interface FooterProps {}
+
+interface DPLFooterProps {
+  className?: string
+}
+
+const DPLFooter: React.FC<DPLFooterProps> = ({ className, children }) => {
+  return (
+    <footer className={`mt-lg ${className}`}>
+      <DPLHr />
+      <div className="py-md ">{children}</div>
+    </footer>
+  )
+}
+
+const DPLFooterSection: React.FC<DPLFooterProps> = ({ children }) => {
+  return <div className="grid justify-items-start gap-8">{children}</div>
+}
 
 const Footer: React.FC<FooterProps> = () => {
   return (
@@ -34,7 +51,6 @@ const Footer: React.FC<FooterProps> = () => {
           </div>
           <ul className="grid grid-flow-col gap-4 text-sm text-gray-400">
             <li>
-              {/* <a href="https://github.com/dev-protocol/niwa/blob/main/TERMS-AND-CONDITIONS.md">Terms and Conditions</a> */}
               <Link to="/terms-and-conditions">Terms and Conditions</Link>
             </li>
             <li>
